requirements.txt
alembic==1.0.8
Mako==1.0.7
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.1
